requirements.txt
streamlit
llama-index
qdrant-client
Pillow